requirements.txt
alembic==1.18.4
asttokens==3.0.1
beautifulsoup4==4.14.3
catboost==1.2.10
certifi==2026.1.4
charset-normalizer==3.4.4
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.61.1
graphviz==0.21
greenlet==3.3.2
idna==3.11
imbalanced-learn==0.14.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
iterative-stratification==0.1.9
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.46.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
narwhals==2.16.0
nest-asyncio==1.6.0
numba==0.64.0
numpy==2.4.2
optuna==4.7.0
packaging==26.0
pandas==3.0.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shap==0.50.0
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.8
soupsieve==2.8.3
SQLAlchemy==2.0.47
stack-data==0.6.3
tabulate==0.9.0
threadpoolctl==3.6.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
xgboost==3.2.0
